Informerly - Day 47
- Adds new loading indicator.
- Adds categoryFeeds model.
- Adds item model.
- Adds menuItems model.
- Adds categories in menu.
- Implements web-services for new categories.
- Testing (In progress)

+++ Informerly/Informerly/CoreData Manager/CoreDataManager.swift
@@ -107,7 +107,7 @@ class CoreDataManager
             println("Error in fetching Bookmarks \(error), \(error?.userInfo)")
         }
         
-        return result;
+        return result.reverse();
     }
     
     class func removeBookmarkFeedOfID(feedID: Int) {
@@ -126,8 +126,10 @@ class CoreDataManager
         var result: Array = managedContext.executeFetchRequest(request, error: &error) as! [BookmarkFeed]
         
         if(error == nil) {
-            managedContext.deleteObject(result[0])
-            managedContext.save(nil)
+            if result.count > 0 {
+                managedContext.deleteObject(result[0])
+                managedContext.save(nil)
+            }
         }
         else {
             println("Error in fetching Bookmarks \(error), \(error?.userInfo)")
